// src/app/models/personal.ts
export class Personal {

    documento: number | string;
    name_lastname: number | string;
    init_date: number | string;
    final_date: number | string;
    time_init_turn: number | string;
    time_final_turn: number | string;
    time_init_extra_hour: number | string;
    time_final_extra_hour: number | string;

}